Prototype
Base scene created to test out some fundamental things we might need in the game.

Includes a script for WASD movement, for having the camera follow the player, and to allow the player to shoot a projectile.

Uses sprites from assests from the asset store. The scripts are designed to make it easy to swap out these sprites as they are only used now as a way to test the scripts.