Small Update
Added the following:
New Sprite for Health/Stamina bar
Script for Health/Stamina Bar
Created an Enemy Health Bar prefab, along with the scripts and sprites for it.
Implemented dashing for the Player (Left Shift). NOTE: The dash effect seems to be affecting world generation script when I changed its layer??
Added our first item, a heal potion. Walk over it to heal.
Reformated the PlayerControl script to inherit off of Matt's Battler Script.

Note: I made changes to the scene in which I added the enemy health bar to the slime gameobject. From now on, I encourage whoever creates enemies to add this prefab to them. Also, combat is not working for now, as well as the animations of the effects are not on the correct layer, and they are not to be used until we make those changes.