Approach and flee states
While the enemy is chasing the player and if it has enough stamina to attack it will enter the approach state(which transitions to the attack state). If the enemy's health is below a predetermined threshold during either the chase or approach states, it will enter the flee state, run away from the player for X amount of time, and enter the idle state.