Minotaur Animations
Ok. Did not know why you really needed me to do anything here if the animations are ready to go.

I am making the assumption that animations are working now because of the  ChangeAnim line i made in Battler. As I said before I needed that to differentiate what walking animations to play. I assume you want to play animations without having to make this function call, which I suggest can be done from making a new trigger or boolean that takes the movement direction and determines whether the battler is moving left or right. Any questions let me know.

Also had to disable and change the minotaur hitboxes to trigger colliders for obvious reasons. And removed the animator handling the hitboxes, as you wanted. Only animations here.